Configurando styles

+++ ticketmaster-frontend/src/components/CategoryList.js
@@ -1,8 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import api from '../services/api';
+import { useNavigate, Link } from 'react-router-dom';
+import '../List.css';
 
 const CategoryList = () => {
   const [categories, setCategories] = useState([]);
+  const navigate = useNavigate();
 
   useEffect(() => {
     api.get('/categories')
@@ -11,19 +14,39 @@ const CategoryList = () => {
   }, []);
 
   const handleDelete = (id) => {
-    api.delete(`/categories/${id}`)
-      .then(() => setCategories(categories.filter(category => category.id !== id)))
-      .catch(error => console.error('Error deleting category:', error));
+    if (window.confirm('Are you sure you want to delete this category?')) {
+      api.delete(`/categories/${id}`)
+        .then(() => {
+          alert('Category deleted successfully.');
+          setCategories(categories.filter(category => category.id !== id))
+        })
+        .catch(error => {
+          console.error('There was an error deleting the category!', error);
+        });
+    }
+  };
+
+  const handleEdit = (id) => {
+    navigate(`/categories/edit/${id}`);
   };
 
   return (
-    <div>
-      <h2>Category List</h2>
-      <ul>
+    <div className="list-container">
+      <div className="list-header">
+        <h2 className="list-title">Categories</h2>
+        <Link to="/categories/new" className="add-button">Create New Category</Link>
+      </div>
+      <ul className="list-group">
         {categories.map(category => (
-          <li key={category.id}>
-            {category.name}
-            <button onClick={() => handleDelete(category.id)}>Delete</button>
+          <li key={category.id} className="list-group-item">
+            <div>
+              <h3 className="item-title">{category.title}</h3>
+              <p className="item-description">{category.name}</p>
+            </div>
+            <div className="action-buttons">
+              <button onClick={() => handleEdit(category.id)} className="edit-btn">Edit</button>
+              <button onClick={() => handleDelete(category.id)} className="delete-btn">Delete</button>
+            </div>
           </li>
         ))}
       </ul>
